Update gh to match my local files
