+++ src/compiler/compat-blocks.js
@@ -1,5 +1,3 @@
-// @ts-check
-
 /**
  * @fileoverview List of blocks to be supported in the compiler compatibility layer.
  * This is only for native blocks. Extensions should not be listed here.
@@ -8,14 +6,6 @@
 // Please keep these lists alphabetical.
 
 const stacked = [
-    // usb to do: compile these when working
-    'camera_movetoxy',
-    'camera_changebyxy',
-    'camera_setx',
-    'camera_changex',
-    'camera_sety',
-    'camera_changey',
-
     'looks_changestretchby',
     'looks_hideallsprites',
     'looks_say',
@@ -28,7 +18,6 @@ const stacked = [
     'motion_glidesecstoxy',
     'motion_glideto',
     'motion_goto',
-    'motion_changebyxy',
     'motion_pointtowards',
     'motion_pointtowardsxy',
     'motion_scroll_right',
@@ -43,9 +32,9 @@ const stacked = [
     'sound_seteffectto',
     'sound_setvolumeto',
     'sound_stopallsounds'
-];
-
-const inputs = [
+  ];
+  
+  const inputs = [
     'looks_effect',
     'motion_xscroll',
     'motion_yscroll',
@@ -53,15 +42,16 @@ const inputs = [
     'sensing_loudness',
     'sensing_userid',
     'sound_volume',
-
+  
     'operator_letter_of',
     'string_item_split',
     'string_convert',
     'string_index_of',
     'string_ternary'
-];
-
-module.exports = {
+  ];
+  
+  module.exports = {
     stacked,
     inputs
-};
+  };
+
